+++ capstone-rs/src/test.rs
@@ -948,12 +948,12 @@ fn test_arch_arm_detail() {
 
     let r0_op_read = ArmOperand {
         op_type: Reg(RegId(ArmReg::ARM_REG_R0 as RegIdInt)),
-        access: Some(RegAccessType::ReadOnly),
+        access: Some(AccessType::ReadOnly),
         ..Default::default()
     };
     let r0_op_write = ArmOperand {
         op_type: Reg(RegId(ArmReg::ARM_REG_R0 as RegIdInt)),
-        access: Some(RegAccessType::WriteOnly),
+        access: Some(AccessType::WriteOnly),
         ..Default::default()
     };
 
@@ -984,7 +984,7 @@ fn test_arch_arm_detail() {
                 &[
                     ArmOperand {
                         op_type: Reg(RegId(ArmReg::ARM_REG_LR as RegIdInt)),
-                        access: Some(RegAccessType::ReadOnly),
+                        access: Some(AccessType::ReadOnly),
                         ..Default::default()
                     },
                     ArmOperand {
@@ -995,7 +995,7 @@ fn test_arch_arm_detail() {
                             disp: -4,
                             lshift: 0,
                         })),
-                        access: Some(RegAccessType::WriteOnly),
+                        access: Some(AccessType::WriteOnly),
                         ..Default::default()
                     },
                 ],
@@ -1013,7 +1013,7 @@ fn test_arch_arm_detail() {
                 &[
                     ArmOperand {
                         op_type: Reg(RegId(ArmReg::ARM_REG_R8 as RegIdInt)),
-                        access: Some(RegAccessType::ReadOnly),
+                        access: Some(AccessType::ReadOnly),
                         ..Default::default()
                     },
                     ArmOperand {
@@ -1024,7 +1024,7 @@ fn test_arch_arm_detail() {
                             disp: -992,
                             lshift: 0,
                         })),
-                        access: Some(RegAccessType::WriteOnly),
+                        access: Some(AccessType::WriteOnly),
                         ..Default::default()
                     },
                 ],
@@ -1087,7 +1087,7 @@ fn test_arch_arm_detail() {
             b"\x70\x47",
             &[ArmOperand {
                 op_type: Reg(RegId(ArmReg::ARM_REG_LR as RegIdInt)),
-                access: Some(RegAccessType::ReadOnly),
+                access: Some(AccessType::ReadOnly),
                 ..Default::default()
             }],
         )],
@@ -1192,12 +1192,12 @@ fn test_arch_arm64_detail() {
                 b"\xbf\x40\x00\xd5",
                 &[
                     Arm64Operand {
-                        access: Some(RegAccessType::WriteOnly),
+                        access: Some(AccessType::WriteOnly),
                         op_type: Pstate(ARM64_PSTATE_SPSEL),
                         ..Default::default()
                     },
                     Arm64Operand {
-                        access: Some(RegAccessType::ReadOnly),
+                        access: Some(AccessType::ReadOnly),
                         op_type: Imm(0),
                         ..Default::default()
                     },
@@ -1209,25 +1209,25 @@ fn test_arch_arm64_detail() {
                 b"\x20\x50\x02\x0e",
                 &[
                     Arm64Operand {
-                        access: Some(RegAccessType::ReadWrite),
+                        access: Some(AccessType::ReadWrite),
                         vas: ARM64_VAS_8B,
                         op_type: Reg(RegId(ARM64_REG_V0 as RegIdInt)),
                         ..Default::default()
                     },
                     Arm64Operand {
-                        access: Some(RegAccessType::ReadOnly),
+                        access: Some(AccessType::ReadOnly),
                         vas: ARM64_VAS_16B,
                         op_type: Reg(RegId(ARM64_REG_V1 as RegIdInt)),
                         ..Default::default()
                     },
                     Arm64Operand {
-                        access: Some(RegAccessType::ReadOnly),
+                        access: Some(AccessType::ReadOnly),
                         vas: ARM64_VAS_16B,
                         op_type: Reg(RegId(ARM64_REG_V2 as RegIdInt)),
                         ..Default::default()
                     },
                     Arm64Operand {
-                        access: Some(RegAccessType::ReadOnly),
+                        access: Some(AccessType::ReadOnly),
                         vas: ARM64_VAS_16B,
                         op_type: Reg(RegId(ARM64_REG_V3 as RegIdInt)),
                         ..Default::default()
@@ -1245,19 +1245,19 @@ fn test_arch_arm64_detail() {
                 b"\x20\xe4\x3d\x0f",
                 &[
                     Arm64Operand {
-                        access: Some(RegAccessType::WriteOnly),
+                        access: Some(AccessType::WriteOnly),
                         vas: ARM64_VAS_2S,
                         op_type: Reg(RegId(ARM64_REG_V0 as RegIdInt)),
                         ..Default::default()
                     },
                     Arm64Operand {
-                        access: Some(RegAccessType::ReadOnly),
+                        access: Some(AccessType::ReadOnly),
                         vas: ARM64_VAS_2S,
                         op_type: Reg(RegId(ARM64_REG_V1 as RegIdInt)),
                         ..Default::default()
                     },
                     Arm64Operand {
-                        access: Some(RegAccessType::ReadOnly),
+                        access: Some(AccessType::ReadOnly),
                         op_type: Imm(3),
                         ..Default::default()
                     },
@@ -1269,15 +1269,15 @@ fn test_arch_arm64_detail() {
                 b"\x00\x18\xa0\x5f",
                 &[
                     Arm64Operand {
-                        access: Some(RegAccessType::ReadWrite),
+                        access: Some(AccessType::ReadWrite),
                         ..s0.clone()
                     },
                     Arm64Operand {
-                        access: Some(RegAccessType::ReadOnly),
+                        access: Some(AccessType::ReadOnly),
                         ..s0
                     },
                     Arm64Operand {
-                        access: Some(RegAccessType::ReadOnly),
+                        access: Some(AccessType::ReadOnly),
                         vector_index: Some(3),
                         op_type: Reg(RegId(ARM64_REG_V0 as RegIdInt)),
                         vas: ARM64_VAS_1S,
@@ -1291,11 +1291,11 @@ fn test_arch_arm64_detail() {
                 b"\xa2\x00\xae\x9e",
                 &[
                     Arm64Operand {
-                        access: Some(RegAccessType::WriteOnly),
+                        access: Some(AccessType::WriteOnly),
                         ..x2.clone()
                     },
                     Arm64Operand {
-                        access: Some(RegAccessType::ReadOnly),
+                        access: Some(AccessType::ReadOnly),
                         vector_index: Some(1),
                         op_type: Reg(RegId(ARM64_REG_V5 as RegIdInt)),
                         vas: ARM64_VAS_INVALID, // should be ARM64_VAS_1D instead?
@@ -1308,7 +1308,7 @@ fn test_arch_arm64_detail() {
                 "dsb",
                 b"\x9f\x37\x03\xd5",
                 &[Arm64Operand {
-                    access: Some(RegAccessType::ReadOnly),
+                    access: Some(AccessType::ReadOnly),
                     op_type: Barrier(Arm64BarrierOp::ARM64_BARRIER_NSH),
                     ..Default::default()
                 }],
@@ -1318,7 +1318,7 @@ fn test_arch_arm64_detail() {
                 "dmb",
                 b"\xbf\x33\x03\xd5",
                 &[Arm64Operand {
-                    access: Some(RegAccessType::ReadOnly),
+                    access: Some(AccessType::ReadOnly),
                     op_type: Barrier(Arm64BarrierOp::ARM64_BARRIER_OSH),
                     ..Default::default()
                 }],
@@ -1331,15 +1331,15 @@ fn test_arch_arm64_detail() {
                 b"\x21\x7c\x02\x9b",
                 &[
                     Arm64Operand {
-                        access: Some(RegAccessType::WriteOnly),
+                        access: Some(AccessType::WriteOnly),
                         ..x1.clone()
                     },
                     Arm64Operand {
-                        access: Some(RegAccessType::ReadOnly),
+                        access: Some(AccessType::ReadOnly),
                         ..x1.clone()
                     },
                     Arm64Operand {
-                        access: Some(RegAccessType::ReadOnly),
+                        access: Some(AccessType::ReadOnly),
                         ..x2.clone()
                     },
                 ],
@@ -1351,17 +1351,17 @@ fn test_arch_arm64_detail() {
                 &[
                     Arm64Operand {
                         // Upstream bug: should be read only, fixed in capstone v6
-                        access: Some(RegAccessType::ReadWrite),
+                        access: Some(AccessType::ReadWrite),
                         op_type: Reg(RegId(ARM64_REG_W1 as RegIdInt)),
                         ..Default::default()
                     },
                     Arm64Operand {
-                        access: Some(RegAccessType::ReadOnly),
+                        access: Some(AccessType::ReadOnly),
                         op_type: Reg(RegId(ARM64_REG_W1 as RegIdInt)),
                         ..Default::default()
                     },
                     Arm64Operand {
-                        access: Some(RegAccessType::ReadOnly),
+                        access: Some(AccessType::ReadOnly),
                         op_type: Imm(0),
                         ..Default::default()
                     },
@@ -1373,17 +1373,17 @@ fn test_arch_arm64_detail() {
                 b"\x00\x40\x21\x4b",
                 &[
                     Arm64Operand {
-                        access: Some(RegAccessType::WriteOnly),
+                        access: Some(AccessType::WriteOnly),
                         op_type: Reg(RegId(ARM64_REG_W0 as RegIdInt)),
                         ..Default::default()
                     },
                     Arm64Operand {
-                        access: Some(RegAccessType::ReadOnly),
+                        access: Some(AccessType::ReadOnly),
                         op_type: Reg(RegId(ARM64_REG_W0 as RegIdInt)),
                         ..Default::default()
                     },
                     Arm64Operand {
-                        access: Some(RegAccessType::ReadOnly),
+                        access: Some(AccessType::ReadOnly),
                         ext: Arm64Extender::ARM64_EXT_UXTW,
                         op_type: Reg(RegId(ARM64_REG_W1 as RegIdInt)),
                         ..Default::default()
@@ -1396,12 +1396,12 @@ fn test_arch_arm64_detail() {
                 b"\xe1\x0b\x40\xb9",
                 &[
                     Arm64Operand {
-                        access: Some(RegAccessType::WriteOnly),
+                        access: Some(AccessType::WriteOnly),
                         op_type: Reg(RegId(ARM64_REG_W1 as RegIdInt)),
                         ..Default::default()
                     },
                     Arm64Operand {
-                        access: Some(RegAccessType::ReadOnly),
+                        access: Some(AccessType::ReadOnly),
                         op_type: Mem(Arm64OpMem(arm64_op_mem {
                             base: ARM64_REG_SP,
                             index: 0,
@@ -1417,11 +1417,11 @@ fn test_arch_arm64_detail() {
                 b"\x20\x04\x81\xda",
                 &[
                     Arm64Operand {
-                        access: Some(RegAccessType::WriteOnly),
+                        access: Some(AccessType::WriteOnly),
                         ..x0.clone()
                     },
                     Arm64Operand {
-                        access: Some(RegAccessType::ReadOnly),
+                        access: Some(AccessType::ReadOnly),
                         ..x1.clone()
                     },
                 ],
@@ -1432,15 +1432,15 @@ fn test_arch_arm64_detail() {
                 b"\x20\x08\x02\x8b",
                 &[
                     Arm64Operand {
-                        access: Some(RegAccessType::WriteOnly),
+                        access: Some(AccessType::WriteOnly),
                         ..x0
                     },
                     Arm64Operand {
-                        access: Some(RegAccessType::ReadOnly),
+                        access: Some(AccessType::ReadOnly),
                         ..x1
                     },
                     Arm64Operand {
-                        access: Some(RegAccessType::ReadOnly),
+                        access: Some(AccessType::ReadOnly),
                         shift: Arm64Shift::Lsl(2),
                         ..x2
                     },
@@ -1452,12 +1452,12 @@ fn test_arch_arm64_detail() {
                 b"\x10\x5b\xe8\x3c",
                 &[
                     Arm64Operand {
-                        access: Some(RegAccessType::WriteOnly),
+                        access: Some(AccessType::WriteOnly),
                         op_type: Reg(RegId(ARM64_REG_Q16 as RegIdInt)),
                         ..Default::default()
                     },
                     Arm64Operand {
-                        access: Some(RegAccessType::ReadOnly),
+                        access: Some(AccessType::ReadOnly),
                         shift: Arm64Shift::Lsl(4),
                         ext: Arm64Extender::ARM64_EXT_UXTW,
                         op_type: Mem(Arm64OpMem(arm64_op_mem {
@@ -3398,13 +3398,13 @@ fn test_arch_x86_detail() {
                 &[
                     X86Operand {
                         size: 2,
-                        access: Some(RegAccessType::WriteOnly),
+                        access: Some(AccessType::WriteOnly),
                         op_type: Reg(RegId(X86_REG_CX as RegIdInt)),
                         ..Default::default()
                     },
                     X86Operand {
                         size: 2,
-                        access: Some(RegAccessType::ReadOnly),
+                        access: Some(AccessType::ReadOnly),
                         op_type: Mem(X86OpMem(x86_op_mem {
                             segment: 0,
                             base: X86_REG_SI,
@@ -3423,7 +3423,7 @@ fn test_arch_x86_detail() {
                 &[
                     X86Operand {
                         size: 1,
-                        access: Some(RegAccessType::ReadWrite),
+                        access: Some(AccessType::ReadWrite),
                         op_type: Mem(X86OpMem(x86_op_mem {
                             segment: 0,
                             base: X86_REG_BX,
@@ -3435,7 +3435,7 @@ fn test_arch_x86_detail() {
                     },
                     X86Operand {
                         size: 1,
-                        access: Some(RegAccessType::ReadOnly),
+                        access: Some(AccessType::ReadOnly),
                         op_type: Reg(RegId(X86_REG_AL as RegIdInt)),
                         ..Default::default()
                     },
@@ -3447,7 +3447,7 @@ fn test_arch_x86_detail() {
                 b"\xd8\x81\xc6\x34",
                 &[X86Operand {
                     size: 4,
-                    access: Some(RegAccessType::ReadOnly),
+                    access: Some(AccessType::ReadOnly),
                     op_type: Mem(X86OpMem(x86_op_mem {
                         segment: 0,
                         base: X86_REG_BX,
@@ -3465,13 +3465,13 @@ fn test_arch_x86_detail() {
                 &[
                     X86Operand {
                         size: 1,
-                        access: Some(RegAccessType::ReadWrite),
+                        access: Some(AccessType::ReadWrite),
                         op_type: Reg(RegId(X86_REG_AL as RegIdInt)),
                         ..Default::default()
                     },
                     X86Operand {
                         size: 1,
-                        access: Some(RegAccessType::ReadOnly),
+                        access: Some(AccessType::ReadOnly),
                         op_type: Mem(X86OpMem(x86_op_mem {
                             segment: 0,
                             base: X86_REG_BX,
@@ -3505,13 +3505,13 @@ fn test_arch_x86_detail() {
                 &[
                     X86Operand {
                         size: 4,
-                        access: Some(RegAccessType::WriteOnly),
+                        access: Some(AccessType::WriteOnly),
                         op_type: Reg(RegId(X86_REG_ECX as RegIdInt)),
                         ..Default::default()
                     },
                     X86Operand {
                         size: 4,
-                        access: Some(RegAccessType::ReadOnly),
+                        access: Some(AccessType::ReadOnly),
                         op_type: Mem(X86OpMem(x86_op_mem {
                             segment: 0,
                             base: X86_REG_EDX,
@@ -3530,13 +3530,13 @@ fn test_arch_x86_detail() {
                 &[
                     X86Operand {
                         size: 4,
-                        access: Some(RegAccessType::ReadWrite),
+                        access: Some(AccessType::ReadWrite),
                         op_type: Reg(RegId(X86_REG_EAX as RegIdInt)),
                         ..Default::default()
                     },
                     X86Operand {
                         size: 4,
-                        access: Some(RegAccessType::ReadOnly),
+                        access: Some(AccessType::ReadOnly),
                         op_type: Reg(RegId(X86_REG_EBX as RegIdInt)),
                         ..Default::default()
                     },
@@ -3549,7 +3549,7 @@ fn test_arch_x86_detail() {
                 &[
                     X86Operand {
                         size: 4,
-                        access: Some(RegAccessType::ReadWrite),
+                        access: Some(AccessType::ReadWrite),
                         op_type: Reg(RegId(X86_REG_ESI as RegIdInt)),
                         ..Default::default()
                     },
@@ -3582,7 +3582,7 @@ fn test_arch_x86_detail() {
                 b"\x55",
                 &[X86Operand {
                     size: 8,
-                    access: Some(RegAccessType::ReadOnly),
+                    access: Some(AccessType::ReadOnly),
                     op_type: Reg(RegId(X86_REG_RBP as RegIdInt)),
                     ..Default::default()
                 }],
@@ -3594,13 +3594,13 @@ fn test_arch_x86_detail() {
                 &[
                     X86Operand {
                         size: 8,
-                        access: Some(RegAccessType::WriteOnly),
+                        access: Some(AccessType::WriteOnly),
                         op_type: Reg(RegId(X86_REG_RAX as RegIdInt)),
                         ..Default::default()
                     },
                     X86Operand {
                         size: 8,
-                        access: Some(RegAccessType::ReadOnly),
+                        access: Some(AccessType::ReadOnly),
                         op_type: Mem(X86OpMem(x86_op_mem {
                             segment: 0,
                             base: X86_REG_RIP,